JK this is actually selection sort.
