+++ src/algorithm/sort/sort.test.ts
@@ -1,10 +1,10 @@
 import { describe, test, expect } from 'vitest';
 import { bubbleSort } from './bubbleSort';
-import { insertionSort } from './insertionSort';
+import { selectionSort } from './selectionSort';
 
 const algorithms: Record<string, AlgorithmWithH<number[]>>  = {
   'Bubble': bubbleSort,
-  'Insertion': insertionSort
+  'Selection': selectionSort
 };
 
 function testAlgorithms(name: string, input: number[], expected: number[]) {
