--- src/algorithm/sort/x.ts
@@ -1,6 +1,0 @@
-import { insertionSort } from "./insertionSort";
-
-const x = [1, 0, -3, 9, 2];
-const y = insertionSort(x);
-
-console.log(y);
